requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
slack_sdk>=3.0.0
openai>=1.0.0
pydantic>=2.0.0
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0
requests>=2.31.0
google-cloud-firestore>=2.11.0
boto3>=1.34.0
anthropic>=0.34.0

